backend: refactor van tracking file structure
Have one class per file and use a more descriptive vantracking module
name.
