+++ backend/src/vantracking/cachetools/cache.py
@@ -4,81 +4,12 @@ from typing import Any, Optional
 
 from cachetools import FIFOCache
 from src.model.stop import Stop
-from src.vans.cache import VanStateCache
-from src.vans.state import Location
+from src.vantracking.cache import VanStateCache
+from src.vantracking.state import Location
+from src.vantracking.cachetools.ttl import TTL
+from src.vantracking.cachetools.ttlqueue import TTLQueue
+from src.vantracking.cachetools.ttlvanstate import TTLVanState
 
-
-class TTL:
-    """
-    Generic wrapper around a time-to-live (TTL) value that will expire after a certain amount of time
-    defined by the class containing an instance of this.
-    """
-
-    def __init__(self, value: Any):
-        # It's best to use monotonic time for this to avoid accidentally expiring items due to
-        # system clock changes.
-        self.timestamp = time.monotonic()
-        self.value = value
-
-    def refresh(self):
-        """
-        Renews the lifespan of this item with a new timestamp.
-        """
-        self.timestamp = time.monotonic()
-
-    def expired(self, ttl: int):
-        """
-        Returns whether this item has expired based on the TTL value specified.
-        """
-        return self.timestamp < time.monotonic() - ttl
-
-
-class TTLQueue:
-    """
-    A queue-like data structure with time-to-live (TTL) functionality. Items are appended to the front, and
-    then popped from the back when they are older than the TTL value specified. Note that the removal of the
-    items only occurs when new items are added, but otherwise any view of the queue will only show items that
-    have not been expired.
-    """
-
-    def __init__(self, ttl: int):
-        self.ttl = ttl
-        self.queue: deque = deque()
-
-    def __contains__(self, key):
-        for item in self.queue:
-            if not item.expired(self.ttl) and item.value == key:
-                return True
-        return False
-
-    def __iter__(self):
-        return map(
-            lambda item: item.value,
-            filter(lambda item: not item.expired(self.ttl), self.queue),
-        )
-
-    def push(self, value: Any):
-        """
-        Pushes a new item to the front of the queue while removing any expired items from the back of the
-        queue.
-        """
-        self.__expire()
-        self.queue.append(TTL(value))
-
-    def __expire(self):
-        while self.queue:
-            item = self.queue[0]
-            if item.expired(self.ttl):
-                self.queue.popleft()
-            else:
-                break
-
-
-class CachedVanState:
-    def __init__(self, locations: TTLQueue, stops: list[Stop], current_stop_index: int):
-        self.locations = locations
-        self.stops = stops
-        self.current_stop_index = current_stop_index
 
 
 class CachetoolsVanStateCache(VanStateCache):
@@ -106,7 +37,7 @@ class CachetoolsVanStateCache(VanStateCache):
         # the cache.
         self.__expire()
         self.cache[van_id] = TTL(
-            CachedVanState(
+            TTLVanState(
                 locations=TTLQueue(ttl=self.ttl),
                 stops=stops,
                 current_stop_index=-1,
