+++ backend/src/vantracking/coordinate.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class Coordinate(BaseModel):
+    latitude: float
+    longitude: float
+
